[CSSSRegions] Incorrect selection clearing on a document without regions https://bugs.webkit.org/show_bug.cgi?id=134901
Reviewed by David Hyatt.

Source/WebCore:

When we select all the content of document with named flows but without regions,
the start and end points of selection is cached in RenderView. However, since
the document has named flows, the selection is split between the subtrees. During the split,
it is possible that the cached end-points of the original selection are not included
in any of the resulting subtree selection and they are not marked accordingly.
In order to process the selection clearing correctly, we have to take the original
selection end-points into account.

Test: fast/regions/selection/select-all-without-regions.html

* rendering/RenderObject.cpp:
(WebCore::RenderObject::isSelectionBorder):

LayoutTests:

* fast/regions/selection/select-all-without-regions-expected.txt: Added.
* fast/regions/selection/select-all-without-regions.html: Added.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@173061 268f45cc-cd09-0410-ab3c-d52691b4dbfc