Apply ARM64-specific lowering to load/store instructions in offlineasm https://bugs.webkit.org/show_bug.cgi?id=136569
Patch by Akos Kiss <akiss@inf.u-szeged.hu> on 2014-09-10
Reviewed by Michael Saboff.

The standard risc lowering of load/store instructions with base +
immediate offset addresses is to move the offset to a temporary, add the
base to the temporary, and then change the load/store to use the
temporary + 0 immediate offset address. However, on ARM64, base +
register offset addressing mode is available, so it is unnecessary to
perform explicit register additions but it is enough to change load/store
to use base + temporary as the address.

* offlineasm/arm64.rb: Added arm64LowerMalformedLoadStoreAddresses

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@173497 268f45cc-cd09-0410-ab3c-d52691b4dbfc