Re-add the request autocomplete feature
<https://bugs.webkit.org/show_bug.cgi?id=136730>

This feature was rolled out in r148731 because it was only used by
Chromium. As we consider supporting this feature, roll it back in, but
leave it disabled.

This rolls out r148731 (which removed the feature) with small changes
Source/JavaScriptCore:

needed to make the code build in ToT, to match modern style, to make
the tests run, and to remove unused code.

Reviewed by Andy Estes.

* Configurations/FeatureDefines.xcconfig:

Source/WebCore:

needed to make the code build in ToT, to match modern style, to make
the tests run, and to remove unused code.

Reviewed by Andy Estes.

Tests: fast/events/constructors/autocomplete-error-event-constructor.html
       fast/forms/form-request-autocomplete.html

* Configurations/FeatureDefines.xcconfig:
* DerivedSources.make:
* WebCore.xcodeproj/project.pbxproj:
* dom/AutocompleteErrorEvent.h: Added.
(WebCore::AutocompleteErrorEvent::create):
(WebCore::AutocompleteErrorEvent::reason):
(WebCore::AutocompleteErrorEvent::AutocompleteErrorEvent):
* dom/AutocompleteErrorEvent.idl: Added.
* dom/EventNames.h:
* dom/EventNames.in:
* html/HTMLAttributeNames.in:
* html/HTMLFormElement.cpp:
(WebCore::HTMLFormElement::HTMLFormElement):
(WebCore::HTMLFormElement::requestAutocomplete):
(WebCore::HTMLFormElement::finishRequestAutocomplete):
(WebCore::HTMLFormElement::requestAutocompleteTimerFired):
(WebCore::HTMLFormElement::parseAttribute):
* html/HTMLFormElement.h:
* html/HTMLFormElement.idl:
* loader/EmptyClients.h:
* loader/FrameLoaderClient.h:
* page/DOMWindow.idl:

Source/WebKit/mac:

needed to make the code build in ToT, to match modern style, to make
the tests run, and to remove unused code.

Reviewed by Andy Estes.

* Configurations/FeatureDefines.xcconfig:
* WebCoreSupport/WebFrameLoaderClient.h:

Source/WebKit2:

needed to make the code build in ToT, to match modern style, to make
the tests run, and to remove unused code.

Reviewed by Andy Estes.

* Configurations/FeatureDefines.xcconfig:
* WebProcess/WebCoreSupport/WebFrameLoaderClient.cpp:
(WebKit::WebFrameLoaderClient::didRequestAutocomplete):
* WebProcess/WebCoreSupport/WebFrameLoaderClient.h:

Source/WTF:

needed to make the code build in ToT, to make the tests run, and to
remove unused code.

Reviewed by Andy Estes.

* wtf/FeatureDefines.h:

LayoutTests:

needed to make the code build in ToT, to match modern style, to make
the tests run, and to remove unused code.

Reviewed by Andy Estes.

* fast/events/constructors/autocomplete-error-event-constructor-expected.txt: Added.
* fast/events/constructors/autocomplete-error-event-constructor.html: Added.
* fast/events/event-creation.html:
* fast/forms/form-request-autocomplete-expected.txt: Added.
* fast/forms/form-request-autocomplete.html: Added.
* js/dom/constructor-length.html:
* platform/efl/js/dom/constructor-length-expected.txt:
* platform/gtk/js/dom/constructor-length-expected.txt:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@173596 268f45cc-cd09-0410-ab3c-d52691b4dbfc