[iOS] Attempt to fix the iOS build after <http://trac.webkit.org/changeset/173606> (https://bugs.webkit.org/show_bug.cgi?id=136789)
Substitute Node::countChildNodes() for Node::childNodeCount() in more places.

Source/WebCore:

Although I updated a comment in WebCore::lastOffsetForEditing() (defined in file
editing/htmlediting.cpp) to refer to Node::countChildNodes(), I'm unclear what the
comment is trying to explain that isn't otherwise explained by the code or the
comments that precede this function. Specifically, I don't understand the use of
the word "preempt" in the comment. We should look to improve the readability of
this comment.

* editing/htmlediting.cpp:
(WebCore::lastOffsetForEditing):
* page/ios/FrameIOS.mm:
(WebCore::Frame::interpretationsForCurrentRoot):

Source/WebKit/mac:

* WebView/WebFrame.mm:
(-[WebFrame focusedNodeHasContent]):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@173608 268f45cc-cd09-0410-ab3c-d52691b4dbfc