REGRESSION (r171197): Clip's content is drawn behind Safari window when entering Full Screen youtube html5 first time https://bugs.webkit.org/show_bug.cgi?id=137029
Reviewed by Brent Fulgham.

In r171197, we ordered the full screen window on- and off-screen so that the fullscreen animation went to the
correct screen, which introduced a subtle glitch at the beginning and end of an animation. For entering, ensure
that the window is scaled to its starting size before ordering on-screen. For exiting, order the window off-
screen before resetting its clipping rect.

* UIProcess/mac/WKFullScreenWindowController.mm:
(-[WKFullScreenWindowController beganEnterFullScreenWithInitialFrame:finalFrame:]):
(-[WKFullScreenWindowController finishedExitFullScreenAnimation:]):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@173888 268f45cc-cd09-0410-ab3c-d52691b4dbfc