Improve the test image diffs page https://bugs.webkit.org/show_bug.cgi?id=137674
Reviewed by Alexey Proskuryakov.

Tools:

Don't treat the input file as a format string, otherwise we have to escape special
characters. Just do a couple of replaces.

* Scripts/webkitpy/layout_tests/controllers/test_result_writer.py:
(TestResultWriter.write_image_diff_files):

LayoutTests:

New version of the image diff page that has the diff images in-place, and allows
for control of the image toggling.

* fast/harness/image-diff-template-expected.txt:
* fast/harness/image-diff-template.html:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@174662 268f45cc-cd09-0410-ab3c-d52691b4dbfc