Add lib directory to mod_perl search path
mod_perl.pl:  Add lib directory to mod_perl search path.  I
accidentally installed updated Perl modules under the Bugzilla
root directory instead of in the system, so this path needed to
be added to make sure up-to-date modules were used.  It may be
removed once Perl modules are updated on the system.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@174849 268f45cc-cd09-0410-ab3c-d52691b4dbfc