[iOS] iPad: Occasional <select> crashes attempting to scroll to non-existing row 0 in viewWillAppear https://bugs.webkit.org/show_bug.cgi?id=138165
Reviewed by David Kilzer.

This is a speculative fix for a crash attempting to scroll to a row in a
select picker on iPad. In these cases we are trying to scroll to the first
row of the first section, but no such row appears to exist. I was unable
to reproduce the issue, but if it is happening we should be able to protect
safely protect against crashing.

* UIProcess/ios/forms/WKFormSelectPopover.mm:
(-[WKSelectTableViewController viewWillAppear:]):
Protect against trying to scroll to a section/row that does not exist
by pre-checking that the section/row is valid.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@175334 268f45cc-cd09-0410-ab3c-d52691b4dbfc