Web Inspector: Lots of TimelineDataGridNode asserts on reloads https://bugs.webkit.org/show_bug.cgi?id=138387
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2014-11-04
Reviewed by Brian Burg.

The TimelineRuler.secondsPerPixel in the TimelineOverview is initialized
with a non-0 value and saved in a setting. However the TimelineRuler in
the OverviewTimelineView is not initialized with this value, and instead
initialized to 0. This patch initializes them to the same non-0 value.

* UserInterface/Views/OverviewTimelineView.js:
(WebInspector.OverviewTimelineView.prototype.set secondsPerPixel):
* UserInterface/Views/TimelineContentView.js:

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@175598 268f45cc-cd09-0410-ab3c-d52691b4dbfc