fast/dom/remove-body-during-body-replacement2.html fails on WK2 https://bugs.webkit.org/show_bug.cgi?id=138334
Reviewed by Anders Carlsson.

Source/WebKit2:

Added an SPI to focus a frame.

* WebProcess/InjectedBundle/API/c/WKBundleFrame.cpp:
(WKBundleFrameFocus):
* WebProcess/InjectedBundle/API/c/WKBundleFramePrivate.h:
* WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::setViewState):

Tools:

Reset focus to main frame between tests. DumpRenderTree does this via -makeFirstResponder
calls, which works because each frame is an NSView in WK1. WK2 makes equivalent
-makeFirstResponder calls, which may or may not be needed for other reasons, but
it also needs to reset internal focus explicitly.

It's not clear if this is the right long-term fix or a workaround for a WebKit bug.
WebKit behavior appears wrong, but it matches Firefox.

* WebKitTestRunner/InjectedBundle/InjectedBundlePage.cpp:
(WTR::InjectedBundlePage::resetAfterTest):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@175638 268f45cc-cd09-0410-ab3c-d52691b4dbfc