Page preview popover should show scaled-down web content https://bugs.webkit.org/show_bug.cgi?id=138444
Reviewed by Anders Carlsson.

Instead of keeping the preferred size in the controller, keep the main view’s 
size, and then scale the bounds of the popover’s view using that and the scale 
factor as a constant. Set the size of the popover explicitly so that it doesn’t 
get the view’s size. This will cause it to scale the WKWebView.
* UIProcess/mac/WKActionMenuController.mm:
(-[WKPagePreviewViewController initWithPageURL:]):
(-[WKPagePreviewViewController loadView]):
(-[WKActionMenuController _createPreviewPopoverForURL:originRect:]):
(-[WKActionMenuController _preferredSizeForPopoverPresentedFromOriginRect:]):



git-svn-id: https://svn.webkit.org/repository/webkit/trunk@175645 268f45cc-cd09-0410-ab3c-d52691b4dbfc