[GTK] Add Since tags to GObject DOM bindings documentation https://bugs.webkit.org/show_bug.cgi?id=139356
Reviewed by Gustavo Noronha Silva.

Now that we have a small stable API, and new symbols are added
manually, we can also add the version to the symbols file, that
the code generator can ue to add Since tags to the gtk-doc.

* bindings/gobject/webkitdom.symbols: Add @2.8 to the new symbols
added for 2.8.
* bindings/scripts/CodeGeneratorGObject.pm:
(GenerateConstants): Add Since tag to gtk-doc if there's a version
number for the symbol in the .symbols file.
(GenerateFunction): Ditto.
(ReadStableSymbols):

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@176920 268f45cc-cd09-0410-ab3c-d52691b4dbfc