Please disable the webkitFirstVersionWithInitConstructorSupport check on Apple TV https://bugs.webkit.org/show_bug.cgi?id=139501
Reviewed by Gavin Barraclough.

Source/JavaScriptCore:

NSVersionOfLinkTimeLibrary only works if you link directly against
JavaScriptCore, which is a bit awkward for our Apple TV client to do.

It's easy enough just to disable this check on Apple TV, since it has no
backwards compatibility requirement.

* API/JSWrapperMap.mm:
(supportsInitMethodConstructors):

Source/WTF:

Added a platform definition for APPLETV.

(It's not an OS because it's not mutually exclusive with other OS definitions.)

* wtf/Platform.h:


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@177091 268f45cc-cd09-0410-ab3c-d52691b4dbfc