HTMLFormControlElement and HTMLObjectElement need not to look for an ancestor on construction https://bugs.webkit.org/show_bug.cgi?id=139791
Patch by Benjamin Poulain <bpoulain@apple.com> on 2014-12-18
Reviewed by Alexey Proskuryakov.

The objects are being constructed, they have not been inserted into the tree yet.
There is not chance of finding the owner form at this point,
that is done when inserting the node into the tree.

* html/HTMLFormControlElement.cpp:
(WebCore::HTMLFormControlElement::HTMLFormControlElement):
* html/HTMLObjectElement.cpp:
(WebCore::HTMLObjectElement::HTMLObjectElement):


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@177538 268f45cc-cd09-0410-ab3c-d52691b4dbfc