Move the CSS Grid properties to the new StyleBuilder https://bugs.webkit.org/show_bug.cgi?id=140051
Reviewed by Sam Weinig.

Move the CSS Grid properties from StyleResolver to the new
StyleBuilder.

This patch adds support for "ConditionalConverter" parameter in
CSSPropertyNames.in to support cases where we only want to call the
property setter conditionally. The converter function is then expected
to return a boolean to indicate if the setter should be called by the
StyleBuilder or not.


git-svn-id: https://svn.webkit.org/repository/webkit/trunk@177872 268f45cc-cd09-0410-ab3c-d52691b4dbfc