Web Inspector: Use spread operator in a few places https://bugs.webkit.org/show_bug.cgi?id=140699
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2015-01-20
Reviewed by Geoffrey Garen.

* UserInterface/Controllers/ProbeManager.js:
(WebInspector.ProbeManager.prototype.get probeSets):
Simplify by using the spread operator.

* UserInterface/Models/TimelineRecording.js:
This actually fixes an issue. The one place this was used was expecting
an array and called ".length" on the result. However, the iterator being
returned does not have a ".length" property so it was always undefined.

git-svn-id: https://svn.webkit.org/repository/webkit/trunk@178787 268f45cc-cd09-0410-ab3c-d52691b4dbfc